:truck: Modify: 스타일컴포넌트 파일 분리
